Doc/elio coding contract/v1/back/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
